:sparkles: [Feat] Review select 분리 및 일부 코드 수정 & aggregate 레포지토리 함수 추가
