+++ src/services/selerts/reviewSelect.ts
@@ -1,0 +1,37 @@
+export const reviewSelect = {
+  id: true,
+  score: true,
+  description: true,
+  createdAt: true,
+  Customer: {
+    select: {
+      User: { select: { name: true } },
+    },
+  },
+}
+
+export const myReviewSelect = {
+  id: true,
+  score: true,
+  description: true,
+  createdAt: true,
+  Mover: {
+    select: {
+      id: true,
+      nickname: true,
+      profileImage: true,
+    },
+  },
+  Estimate: {
+    select: {
+      isAssigned: true,
+      price: true,
+      MovingInfo: {
+        select: {
+          movingDate: true,
+          movingType: true,
+        },
+      },
+    },
+  },
+};
