--- src/services/selerts/reviewSelert.ts
@@ -1,5 +1,0 @@
-export const reviewSelect = {
-  id: true,
-  score: true,
-  description: true,
-};
